ExprEval

Description

Documentation

data EvalResult a

EvalResult is a type that holds the result of the evaluation of an expression

Constructors

EvalError String
Wraps an error string if there is an error during computation

Result a
Wraps the result of a successful computation

isInfinity :: AllNums a => a -> Bool

isInfinity takes a Double, Float, Int or Integer and returns a Boolean result

Class Declaration

class ExprEval a where

Minimal complete definition

eval, simplify

Methods

eval :: Map String a -> Expr a -> EvalResult a

simplify :: Map String a -> Expr a -> Expr a

(!+) :: AllNums a => Expr a -> Expr a -> Expr a

(!-) :: AllNums a => Expr a -> Expr a -> Expr a

(!*) :: AllNums a => Expr a -> Expr a -> Expr a

(!/) :: AllNums a => Expr a -> Expr a -> Expr a

nPow :: AllNums a => Expr a -> Expr a

log' :: AllNums a => a -> Expr a -> Expr a

ln :: AllNums a => Expr a -> Expr a

sine :: AllNums a => Expr a -> Expr a

cosine :: AllNums a => Expr a -> Expr a

(!^) :: AllNums a => Expr a -> Expr a -> Expr a

val :: AllNums a => a -> Expr a

var :: AllNums a => String -> Expr a

class (Num a, Eq a, Show a, Ord a) => AllNums a where

Allows eval and simplify to be used with Double, Float, Integer and Int types by creating instances for each type and defining mathematical operators in them

Minimal complete definition

numDiv, numE, numLog, numLn, numCos, numSin, numPow

Methods

numDiv :: a -> a -> a

numE :: a -> a

numLog :: a -> a -> a

numLn :: a -> a

numCos :: a -> a

numSin :: a -> a

numPow :: a -> a -> a