ExprTest

Description

Documentation

simplifyProp1 :: Expr Double -> Bool

Simplifying an expression once should be the same result as simplifying it multiple times

simplifyProp2 :: Double -> Double -> String -> Bool

if the variable is not in the list of vars, it should be returned as a variable

simplifyProp3 :: [Integer] -> Bool

An expression that is of the form Add (Const a) (Add (Const b) (Add (etc.) (etc.))) should simplify to a number

listToExpr :: [Integer] -> Expr Integer

sampleExpr1 :: [Char]

Orphan instances

Arbitrary (Expr Double)

Methods

arbitrary :: Gen (Expr Double)

shrink :: Expr Double -> [Expr Double]